Image_From_Text_Generator_StableDiffusion/requirements.txt
customtkinter==4.6.1
darkdetect==0.7.1
diffusers
filelock==3.8.0
huggingface-hub==0.9.1
idna==3.4
importlib-metadata==4.12.0
numpy
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
tk==0.1.0
tokenizers==0.12.1 
torch
torchaudio
torchvision
tqdm
transformers
typing_extensions
urllib3
zipp